Fix ima/evm digital signature reboot failed issue

+++ tests/security/ima/ima_appraisal_audit.pm
@@ -1,4 +1,4 @@
-# Copyright (C) 2019 SUSE LLC
+# Copyright (C) 2019-2021 SUSE LLC
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -16,7 +16,7 @@
 # Summary: Test audit function for IMA appraisal
 # Note: This case should come after 'ima_appraisal_digital_signatures'
 # Maintainer: llzhao <llzhao@suse.com>
-# Tags: poo#49568
+# Tags: poo#49568, poo#92347
 
 use base "opensusebasetest";
 use strict;
@@ -37,29 +37,33 @@ sub run {
     my $sample_cmd = 'yes --version';
 
     systemctl('is-active auditd');
+    if (script_run("grep CONFIG_INTEGRITY_TRUSTED_KEYRING=y /boot/config-`uname -r`") == 0) {
+        record_soft_failure("bsc#1157432 for SLE15SP2+: CA could not be loaded into the .ima or .evm keyring");
+    }
+    else {
+        # Make sure IMA is in the enforce mode
+        validate_script_output "grep -E 'ima_appraise=(fix|log|off)' /etc/default/grub || echo 'IMA enforced'", sub { m/IMA enforced/ };
+        assert_script_run("test -e /etc/sysconfig/ima-policy", fail_message => 'ima-policy file is missing');
 
-    # Make sure IMA is in the enforce mode
-    validate_script_output "grep -E 'ima_appraise=(fix|log|off)' /etc/default/grub || echo 'IMA enforced'", sub { m/IMA enforced/ };
-    assert_script_run("test -e /etc/sysconfig/ima-policy", fail_message => 'ima-policy file is missing');
+        # Clear security.ima no matter whether it existing
+        script_run "setfattr -x security.ima $sample_app";
+        validate_script_output "getfattr -m security.ima -d $sample_app", sub { m/^$/ };
 
-    # Clear security.ima no matter whether it existing
-    script_run "setfattr -x security.ima $sample_app";
-    validate_script_output "getfattr -m security.ima -d $sample_app", sub { m/^$/ };
+        assert_script_run("echo -n '' > /var/log/audit/audit.log");
+        my $ret = script_output($sample_cmd, 30, proceed_on_failure => 1);
+        die "$sample_app should not have permission to run" if ($ret !~ "\Q$sample_app\E: *Permission denied");
+        validate_script_output "ausearch -m INTEGRITY_DATA", sub { m/\Q$sample_app\E/ };
 
-    assert_script_run("echo -n '' > /var/log/audit/audit.log");
-    my $ret = script_output($sample_cmd, 30, proceed_on_failure => 1);
-    die "$sample_app should not have permission to run" if ($ret !~ "\Q$sample_app\E: *Permission denied");
-    validate_script_output "ausearch -m INTEGRITY_DATA", sub { m/\Q$sample_app\E/ };
+        # Test both default(no ima_apprais=) and ima_appraise=log situation
+        add_grub_cmdline_settings("ima_appraise=log", update_grub => 1);
+        power_action('reboot', textmode => 1);
+        $self->wait_boot(textmode => 1);
+        $self->select_serial_terminal;
 
-    # Test both default(no ima_apprais=) and ima_appraise=log situation
-    add_grub_cmdline_settings("ima_appraise=log", update_grub => 1);
-    power_action('reboot', textmode => 1);
-    $self->wait_boot(textmode => 1);
-    $self->select_serial_terminal;
-
-    assert_script_run("echo -n '' > /var/log/audit/audit.log");
-    assert_script_run "$sample_cmd";
-    validate_script_output "ausearch -m INTEGRITY_DATA", sub { m/\Q$sample_app\E/ };
+        assert_script_run("echo -n '' > /var/log/audit/audit.log");
+        assert_script_run "$sample_cmd";
+        validate_script_output "ausearch -m INTEGRITY_DATA", sub { m/\Q$sample_app\E/ };
+    }
 }
 
 sub test_flags {
