+++ tests/security/ima/ima_appraisal_digital_signatures.pm
@@ -1,4 +1,4 @@
-# Copyright (C) 2019 SUSE LLC
+# Copyright (C) 2019-2021 SUSE LLC
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -15,7 +15,7 @@
 #
 # Summary: Test IMA appraisal using digital signatures
 # Maintainer: llzhao <llzhao@suse.com>
-# Tags: poo#49154
+# Tags: poo#49154, poo#92347
 
 use base "opensusebasetest";
 use strict;
@@ -64,26 +64,32 @@ sub run {
         m/security\.ima=[0-9a-zA-Z+\/]{355}/;
     };
 
-    # Prepare mok ceritificate file
-    assert_script_run "mkdir -p /etc/keys/ima";
-    assert_script_run "cp $cert_der /etc/keys/ima/";
+    if (script_run("grep CONFIG_INTEGRITY_TRUSTED_KEYRING=y /boot/config-`uname -r`") == 0) {
+        record_soft_failure("bsc#1157432 for SLE15SP2+: CA could not be loaded into the .ima or .evm keyring");
+    }
+    else {
+        # Prepare mok ceritificate file
+        assert_script_run "mkdir -p /etc/keys/ima";
+        assert_script_run "cp $cert_der /etc/keys/ima/";
 
-    assert_script_run "wget --quiet " . data_url("ima/ima_appraisal_ds_policy" . " -O /etc/sysconfig/ima-policy");
+        assert_script_run "wget --quiet " . data_url("ima/ima_appraisal_ds_policy" . " -O /etc/sysconfig/ima-policy");
 
-    replace_grub_cmdline_settings('ima_appraise=fix', '', update_grub => 1);
+        replace_grub_cmdline_settings('ima_appraise=fix', '', update_grub => 1);
 
-    power_action('reboot', textmode => 1);
-    $self->wait_boot(textmode => 1);
-    $self->select_serial_terminal;
+        power_action('reboot', textmode => 1);
+        $self->wait_boot(textmode => 1);
+        $self->select_serial_terminal;
 
-    assert_script_run "dmesg | grep IMA:.*completed";
+        assert_script_run "dmesg | grep IMA:.*completed";
 
-    # Remove security.ima attribute manually, and verify it is empty
-    assert_script_run "setfattr -x security.ima $sample_app";
-    validate_script_output "getfattr -m security.ima -d $sample_app", sub { m/^$/ };
+        # Remove security.ima attribute manually, and verify it is empty
+        assert_script_run "setfattr -x security.ima $sample_app";
+        validate_script_output "getfattr -m security.ima -d $sample_app", sub { m/^$/ };
 
-    my $ret = script_output($sample_cmd, 30, proceed_on_failure => 1);
-    die "$sample_app should not have permission to run" if ($ret !~ "\Q$sample_app\E: *Permission denied");
+        my $ret = script_output($sample_cmd, 30, proceed_on_failure => 1);
+        die "$sample_app should not have permission to run" if ($ret !~ "\Q$sample_app\E: *Permission denied");
+
+    }
 }
 
 sub test_flags {
